refactor(building-components): rename lod_host to LodHostHelper
Clarify that the module only builds posed level-root kit content, not
`LodSceneHost` scaffolding (that lives in `lod::lod_host_scene_pending` /
`LodScene::host`).

## Agent Dialogue

- **Developer:** After the chunk/Avian buildings LOD migration, ask whether
  `lod_host.rs` is still needed; then rename it to `LodHostHelper` for clarity.
- **Agent:** Kept the posed-kit helpers (still used by partitions/joints);
  renamed module to `lod_host_helper` with unit struct `LodHostHelper::posed_*`.
