+++ maybraid/richmond/building-components/src/partitions.rs
@@ -4,7 +4,7 @@
 //! [`PartitionNode`] covers both **direct** kit mappings and **tessellated** forms (polyline / arc).
 //! Primitive kit geometry — higher-order walls/portals live in `richmond_buildings::{paneling,arcs,portals}`.
 //!
-//! Host split: crate [`crate::lod_host`] (structure) vs [`host`] (partition GLB resolution policy).
+//! Host split: crate [`crate::LodHostHelper`] (posed content) vs [`host`] (partition GLB resolution policy).
 
 pub mod geometry;
 pub mod host;
